+++ packages/govuk-prototype-components/package.json
@@ -12,6 +12,9 @@
   "exports": {
     ".": {
       "import": "./x-govuk/all.js"
+    },
+    "./decorate": {
+      "import": "./lib/decorate.js"
     }
   },
   "repository": {
