+++ packages/govuk-prototype-rig/lib/globals/index.js
@@ -1,1 +1,7 @@
-export { helperGlobals as globals } from './helper.js'
+import { decorate } from 'govuk-prototype-components/decorate'
+import { checked } from './checked.js'
+
+export const globals = {
+  checked,
+  decorate
+}
